Revert "[WIP] iommu/amd: Include PCI segment ID when initialize IOMMU."
This reverts commit a9a596c8acb7e80383b2dcfc51018d76116321f6.

+++ drivers/iommu/amd/amd_iommu_types.h
@@ -735,8 +735,8 @@ struct acpihid_map_entry {
 	struct list_head list;
 	u8 uid[ACPIHID_UID_LEN];
 	u8 hid[ACPIHID_HID_LEN];
-	u32 devid;
-	u32 root_devid;
+	u16 devid;
+	u16 root_devid;
 	bool cmd_line;
 	struct iommu_group *group;
 };
@@ -744,7 +744,7 @@ struct acpihid_map_entry {
 struct devid_map {
 	struct list_head list;
 	u8 id;
-	u32 devid;
+	u16 devid;
 	bool cmd_line;
 };
